Remove type information from Interceptor + Operation classes (#6)
This is a heavy restructure of the library. The primary goal is to remove type information from the Interceptor and Operation classes. This is because a list of Interceptors is created and then reused for many different input/output types, so type information inside an Interceptor isn't useful or truthful.

- Remove type params from Interceptor and Operation classes
- Replace InterceptorChain with OperationPipeline
- Operation class now implements two interfaces: ExecutableOperation (used by OperationPipeline creators) and ContinuableOperation (used by Interceptors). This makes it more clear what should be called in either scenario